+++ package.json
@@ -35,12 +35,10 @@
     "sinon-chai": "2.8.0"
   },
   "dependencies": {
-    "async": "^1.5.0",
+    "async": "1.5.0",
     "eslint": "1.8.0",
     "eslint-plugin-react": "3.7.1",
     "glob-all": "3.0.1",
     "minimist": "1.1.1",
-    "moment": "2.10.6",
-    "snyk-report": "git+ssh://git@github.com/holidayextras/snyk-report.git#exports"
-  }
+    "moment": "2.10.6"  }
 }
